+++ NEWS.md
@@ -1,3 +1,9 @@
+## package inpdfr v0.1.8
+* fix calling if() with a vector of length 2 or more
+
+## package inpdfr v0.1.7 
+* fix calling if() with a vector of length 2 or more
+
 ## package inpdfr v0.1.6 (bug-fix release)
 * fix WARN "working directory change"
 * GUI using RGtk2 no longer part of the package